requirements.txt
streamlit==1.32.0
tensorflow-cpu==2.15.0
tf_keras
h5py
numpy<2.0.0
Pillow
pandas